refactor: [US-058] - Move enrichment tasks into core/tasks/enrichment.py (wire names pinned)
